Add files via upload
The map shows circles with radii dictated by the number of ride-sharing pickups to occur at a specific GPS location. The circles include the location as well as the most frequent hour that pickups occurred at that location (where warm colors constitute a gradient for day pickups and cool colors a likewise gradient for night pickups).